refactored lib/ to  subdirs
