+++ lib/main.dart
@@ -1,10 +1,10 @@
 import 'package:flutter/material.dart';
 import 'package:firebase_core/firebase_core.dart';
 import 'package:provider/provider.dart';
-import 'package:realty_group_project/UserModel.dart';
-import 'ListingsPage.dart';
-import 'StarredListings.dart';
-import 'firebase_options.dart';
+import 'package:realty_group_project/models/UserModel.dart';
+import 'pages/ListingsPage.dart';
+import 'widgets/StarredListings.dart';
+import 'services/firebase_options.dart';
 
 void main() {
   runApp(
