+++ lib/pages/SearchHistory.dart
